backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
twilio
python-dotenv
